% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{gm_dimensions}
\alias{gm_dimensions}
\title{gm_dimensions}
\usage{
gm_dimensions(x)
}
\value{
A list containing
\itemize{
\item{dimensions - a 5-column matrix where \itemize{
\item{column 1 - the start-index of the geometry}
\item{column 2 - the end-index of the geometry}
\item{column 3 - the dimension of the geometry}
\item{column 4 - the nesting level of the geometry}
\item{column 5 - the SEXP type of the geometry}
}
}
\item{max_dimension - the maximum dimension in all the geometries}
\item{max_nest - the maximum nesting level in all the geometries}
}
}
\description{
Calculates the dimensions of geometries.
}
\examples{

## A vector (representing an POINT)
x <- 1:2
gm_dimensions( x )

## A matrix (representing a LINE)
m <- matrix(1:6, ncol = 2 )
gm_dimensions( m )

## A list (representing a POLYGON)
l <- list( m )
gm_dimensions( l )

## A list of lists (representing a MULTIPOLYGON)
l <- list( l )
gm_dimensions( l )

## A list of matrices, points and lists (representing multiple, mixed geometries)
l <- list( x, m, list( m ) )
gm_dimensions( l )

}
