% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometries.R
\name{gm_make_geometries}
\alias{gm_make_geometries}
\title{make geometries}
\usage{
gm_make_geometries(df, id_cols, geometry_cols, class_attributes = NULL)
}
\arguments{
\item{df}{data.frame}

\item{id_cols}{vector of id columns (iether integer or string)}

\item{geometry_cols}{vector of geometry columns (either integer or string)}

\item{class_attributes}{class attributes to assign to each geometry}
}
\description{
Converts a data.frame into a collection of geometries.
}
\examples{

df <- data.frame(
  id = c(1,1,1,1,1,2,2,2,2,2)
  , x = 1:10
  , y = 10:1
)

gm_make_geometries(df, id_cols = c(1), geometry_cols = c(2,3))

## Adding a class attribute
gm_make_geometries(df, id_cols = c(1), geometry_cols = c(2,3), "my_line_object")

## Adding a second ID column
df$id1 <- c(1,1,1,2,2,1,1,2,2,3)
gm_make_geometries(df, id_cols = c(1,4), geometry_cols = c(2,3), "my_multiline_object")

## Using character column names
gm_make_geometries(df, id_cols = c("id","id1"), geometry_cols = c("x","y") )

}
